Updates in makefile

+++ ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: mgumienn <mgumienn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/23 08:56:19 by marvin            #+#    #+#             */
-/*   Updated: 2025/09/23 18:57:44 by mgumienn         ###   ########.fr       */
+/*   Updated: 2025/09/24 16:18:24 by mgumienn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_bzero(void *ptr, unsigned int n)
+#include "libft.h"
+
+void	*ft_bzero(void *ptr, size_t n)
 {
 	unsigned char	*p;
