+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: mgumienn <mgumienn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/22 14:46:47 by mgumienn          #+#    #+#             */
-/*   Updated: 2025/09/23 18:50:56 by mgumienn         ###   ########.fr       */
+/*   Updated: 2025/09/24 19:35:20 by mgumienn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *ptr, unsigned int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *ptr, unsigned int c, size_t n)
 {
 	unsigned char	*p;
